// src/Client/src/app/models/area.ts
export class Area {
    REG?: string;
    REG_NAME?: string;
    CWT?: string;
    CWT_NAME?: string;
    AMP?: string;
    AMP_NAME?: string;
    TAM?: string;
    TAM_NAME?: string;
    DISTRICT?: string;
    MUN?:string;
    MUN_NAME?: string;
    TAO?: string;
    TAO_NAME?: string;
    EA?: string;
    VIL?: string;
    VIL_NAME?: string;
    MAP_STATUS?: string;
    Building?: number;
    Household?: number;
    Agricultural_HH?: number;
    ES_BUSI?: number;
    ES_INDUS?: number;
    ES_HOTEL?: number;
    ES_PV_HOS?: number;
    REMARK?: string;
    FI?:string[];
    FI_H?:string[];
    FS?:string;
    FS_H?:string[];
}